Adding scalastyle
