+++ core/src/main/scala/iliad/data.scala
@@ -1,9 +1,9 @@
 package iliad
 
-import iliad.kernel.GL
+import iliad.kernel._
 
 import java.nio.ByteBuffer
 
 case class VertexData(data: ByteBuffer, numVertices: Int)
 case class ElementData(data: ByteBuffer, numElements: Int)  
-case class ModelData(buffer: GL.BufferInstance, vertexData: VertexData, elementData: ElementData)
+case class ModelData(buffer: BufferInstance, vertexData: VertexData, elementData: ElementData)
